Big overhaul
- Identify as KetchupBot-Updater to Galaxy Info API
- Galaxy Info API endpoint is now public, so no need for authentication
- Get rid of ShipData and simply use Dictionaries instead
- Move away from using Task.WhenAll to Parallel.ForEachAsync
  - Honestly.. I still don't understand the difference, or which one to
use. I asked GitHub Copilot and it suggested using Parallel.ForEachAsync
for some reason? Probably does some better task management under the
hood or something. Whatever the reason is, it still seems to work, and
allowed me to do the following
- Add MassUpdateShips to ShipUpdater, and move the bulk of the mass
updating code there. UpdateAllShips now serves to just grab a lists of
every ship, then instruct MassUpdateShips to update each ship using the
already gathered data
- Make _cachedshipdata use the serialized version, rather than the
string version
- Move the function wide profiling code outside of UpdateShip, since it
tended to be rather inaccurate.
  - With C#'s multithreaded design, certain parts of async functions can
run on numerous different threads. This essentially means that at the
start of the mass processing, a thread would literally start the
stopwatch for EVERY ship in the processing queue, but leave the bulk of
the processing for the other threads. This means that the start would be
at the start of the mass processing, not the start of the individual
ship's processing.
- Add a ton of docs to functions everywhere

While I haven't necessarily tested this stuff (since the galaxy info api
is down), it should be a good place to make the switch. The foundation
is there, and it's a lot easier to maintain. I can probably add in the
rest of the features soon